Rename fraction-indicator directory to proportion-indicator
