+++ src/shared/components/proportion-indicator/CircularProportionIndicator.scss
