+++ src/shared/components/proportion-indicator/CircularProportionIndicator.tsx
