+++ src/shared/components/proportion-indicator/LinearProportionIndicator.scss
