+++ src/shared/components/proportion-indicator/LinearProportionIndicator.tsx
