+++ stories/shared-components/proportion-indicator/ProportionIndicator.stories.scss
